add new docker_base.py for simplification of docker implementations cleanup code
