+++ utils/servers/base.py
@@ -1,8 +1,11 @@
 import asyncio
 import datetime
 import logging
+from typing import List, Tuple
 
 import hikari
+import lightbulb
+import regex
 
 
 class BaseServer:
@@ -44,6 +47,27 @@ class BaseServer:
 
     async def update_server_information(self):
         await self.bot.add_game_presence(self.name, self.name)
+
+    def check_for_mentions(self, message: str) -> Tuple[List[hikari.snowflakes.Snowflakeish], str]:
+        indexes: List[int] = [m.start() for m in regex.finditer('@', message)]
+        mentioned_members = []
+        for index in indexes:
+            try:
+                mention = message[index + 1:]
+                for chan in self.bot.chat_channels_obj:
+                    for ind in range(0, min(len(mention) + 1, 32)):
+                        member = lightbulb.utils.find(self.bot.cache.get_guild(chan.guild_id).get_members().values(),
+                                                      lambda m: m.username == mention[:ind] or
+                                                                m.nickname == mention[:ind])
+                        if member:
+                            mentioned_members.append(member)
+                            message = message.replace("@" + mention[:ind], f"<@{member.id}>")
+                            break
+
+            except Exception as e:
+                logging.critical("ERROR | Server2Guild Mentions Exception caught: " + str(e))
+                pass
+        return mentioned_members, message
 
     @staticmethod
     async def sleep_with_backoff(tries, wait_time=5):
